requirements.txt
Django>=5.1.2
Pillow>=10.0.0
googletrans==4.0.0rc1
polib>=1.2.0
django-ratelimit>=4.1.0
deep-translator>=1.11.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.9
google-api-python-client>=2.100.0
google-auth>=2.22.0
requests>=2.31.0
